requirements.txt
aioredis==2.0.1
anyio==3.6.2
APScheduler==3.10.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
deepdiff==6.2.3
distlib==0.3.6
ecdsa==0.18.0
fastapi==0.92.0
filelock==3.9.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
loguru==0.6.0
MarkupSafe==2.1.2
ordered-set==4.1.0
orjson==3.8.7
pipenv==2023.2.18
platformdirs==3.0.0
pyasn1==0.4.8
pydantic==1.10.5
PyMySQL==1.0.2
python-jose==3.3.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
